pyproject.toml
[tool.poetry]
name = "palace-diffusion"
version = "0.1.0"
description = ""
authors = ["Adrienne Wright <ad@rienne.de>"]
readme = "README.md"
packages = [{include = "palace_diffusion"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.5"
python-dotenv = "^1.0.0"
transformers = "^4.33.0"
torch = ">=2.0.0, !=2.0.1"
black = "^23.7.0"
accelerate = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
